% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_gpdd.R
\name{download_gpdd}
\alias{download_gpdd}
\title{Download and Save Data from GPDD.}
\usage{
download_gpdd(dataset_name = c("gpdd_data", "gpdd_main",
  "gpdd_timeperiod", "gpdd_taxon", "gpdd_datasource", "gpdd_biotope",
  "gpdd_location"), dir = rappdirs::user_data_dir("gpdd"),
  overwrite = "Ask")
}
\arguments{
\item{dataset_name}{A string of the name of the dataset.}

\item{dir}{A directory to which the datasets will be downloaded.}

\item{overwrite}{Indicates whether to overwrite existing datasets or not.}
}
\value{
The requested dataset in tibble format.
}
\description{
Download and Save Data from GPDD.
}
\examples{
download_gpdd()
download_gpdd(c("gpdd_data", "gpdd_timeperiod"))
}
\author{
Zihao Li, \email{zihao0104@berkeley.edu}
}
