% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpdd.R
\name{gpdd}
\alias{gpdd}
\title{Download Tables from the GPDD Data}
\usage{
gpdd(table = c("data", "main", "timeperiod", "location"))
}
\arguments{
\item{table}{name of the table to be downloaded.}
}
\value{
the requested data.frame (as a tibble)
}
\description{
Download Tables from the GPDD Data
}
\examples{
gpdd("location")
}
