% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_gpdd.R
\name{load_gpdd}
\alias{load_gpdd}
\title{Load GPDD data from local directory.}
\usage{
load_gpdd(dataset_name = c("gpdd_data", "gpdd_main", "gpdd_timeperiod",
  "gpdd_taxon", "gpdd_datasource", "gpdd_biotope", "gpdd_location"),
  dir = rappdirs::user_data_dir())
}
\arguments{
\item{dataset_name}{A string of the name of the dataset.}

\item{dir}{A directory from which the dataset will be loaded.}
}
\value{
The requested dataset loaded in tibble format.
}
\description{
Load GPDD data from local directory.
}
\examples{
load_gpdd()
load_gpdd("gpdd_data")
load_gpdd(c("gpdd_main", "gpdd_timeperiod"))
}
\author{
Zihao Li, \email{zihao0104@berkeley.edu}
}
