% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_gpdd.R
\name{load_gpdd}
\alias{load_gpdd}
\title{Load GPDD data from local directory.}
\usage{
load_gpdd(dataset_name = c("data", "main", "timeperiod", "taxon",
  "datasource", "biotope", "location"),
  dir = rappdirs::user_data_dir("gpdd"))
}
\arguments{
\item{dataset_name}{A string of the name of the dataset.}

\item{dir}{A directory from which the dataset will be loaded.}
}
\value{
The requested dataset loaded in tibble format.
}
\description{
Load GPDD data from local directory.
}
\examples{
load_gpdd()
load_gpdd("data")
load_gpdd(c("main", "timeperiod"))
}
\author{
Zihao Li, \email{zihao0104@berkeley.edu}
}
